+++ src/main/java/com/activiti/extension/config/KeyCloakPropertyLoader.java
@@ -1,4 +1,4 @@
-package com.activiti.extension.bean;
+package com.activiti.extension.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
